+++ nars_gui/src/main/java/automenta/vivisect/surfaceplotter/surface/Projector.java
@@ -392,8 +392,8 @@ float zfactor;
     // elevates and projects
 
     temp = factor / (y * cos_elevation - z * sz_sin + distance);
-    return new Point((int)(Math.round(x * temp) + trans_x),
-                     (int)(Math.round((y * sin_elevation + z * sz_cos) * -temp) + trans_y));
+    return new Point(Math.round(x * temp) + trans_x,
+            Math.round((y * sin_elevation + z * sz_cos) * -temp) + trans_y);
   }
