+++ nars_gui/src/main/java/automenta/vivisect/swing/NSlider.java
@@ -89,7 +89,7 @@ public class NSlider extends JLabel implements MouseListener, MouseMotionListene
         if (barColor == null) {
             //Green->Yellow->Red
             //g.setColor(Color.getHSBColor( (1f - (float)p) / 3.0f , 0.2f, 0.9f));
-             g.setColor(Color.getHSBColor( (1f - (float)p) / 3.0f , 0.2f, 0.8f + 0.15f));
+             g.setColor(Color.getHSBColor( (1f - p) / 3.0f , 0.2f, 0.8f + 0.15f));
             
         }
         else {
