+++ nars_gui/src/main/java/nars/gui/NARControls.java
@@ -877,7 +877,7 @@ public class NARControls extends JPanel implements ActionListener, Reaction {
 
             @Override
             public void setValue(float v) {
-                int i = (int) Math.round(v);
+                int i = Math.round(v);
                 super.setValue(i);
                 x.set(i);
             }
