+++ nars_gui/src/main/java/nars/gui/output/graph/deprecated/ProcessingGraphPanel2.java
@@ -114,7 +114,7 @@ abstract public class ProcessingGraphPanel2<V,E> extends NPanel {
         NSlider edgeDist = new NSlider(edgeDistance, 1, 100) {
             @Override
             public void onChange(float v) {
-                edgeDistance = (float)v;
+                edgeDistance = v;
                 ProcessingGraphPanel2.this.update();
             }          
         };
